@@ -7,7 +7,8 @@
 #include "ORB_SLAM2/Error.h"
 #include "ORB_SLAM2/ORBExtractor.h"
 
-namespace ORB_SLAM2_ROS2 {
+namespace ORB_SLAM2_ROS2
+{
 
 /**
  * @brief 四叉树根节点构造函数
@@ -20,9 +21,10 @@ QuadtreeNode::QuadtreeNode(int x, int y, const std::vector<cv::KeyPoint> &keyPoi
     , mfRowEnd(y)
     , mfColBegin(0)
     , mfColEnd(x)
-    , mpAllKeyPoints(&keyPoints) {
-    for (std::size_t i = 0; i < keyPoints.size(); ++i)
-        mvKeyPoints.push_back(i);
+    , mpAllKeyPoints(&keyPoints)
+{
+  for (std::size_t i = 0; i < keyPoints.size(); ++i)
+    mvKeyPoints.push_back(i);
 }
 
 /**
@@ -39,13 +41,15 @@ QuadtreeNode::QuadtreeNode(SharedPtr pParent, double rowBegin, double rowEnd, do
     , mfRowBegin(rowBegin)
     , mfRowEnd(rowEnd)
     , mfColBegin(colBegin)
-    , mfColEnd(colEnd) {
-    mpAllKeyPoints = pParent->mpAllKeyPoints;
-    for (auto &idx : pParent->mvKeyPoints) {
-        const auto &kp = mpAllKeyPoints->at(idx);
-        if (isIn(kp))
-            mvKeyPoints.push_back(idx);
-    }
+    , mfColEnd(colEnd)
+{
+  mpAllKeyPoints = pParent->mpAllKeyPoints;
+  for (auto &idx : pParent->mvKeyPoints)
+  {
+    const auto &kp = mpAllKeyPoints->at(idx);
+    if (isIn(kp))
+      mvKeyPoints.push_back(idx);
+  }
 }
 
 /**
@@ -53,15 +57,18 @@ QuadtreeNode::QuadtreeNode(SharedPtr pParent, double rowBegin, double rowEnd, do
  *
  * @param pParent 指定父节点信息
  */
-void QuadtreeNode::split(QuadtreeNode::SharedPtr pParent) {
-    std::vector<double> rows = {mfRowBegin, (mfRowBegin + mfRowEnd) / 2, mfRowEnd};
-    std::vector<double> cols = {mfColBegin, (mfColBegin + mfColEnd) / 2, mfColEnd};
-    for (int i = 0; i < 2; ++i) {
-        for (int j = 0; j < 2; ++j) {
-            auto node = std::make_shared<QuadtreeNode>(pParent, rows[i], rows[i + 1], cols[j], cols[j + 1]);
-            mvpChildren.push_back(node);
-        }
+void QuadtreeNode::split(QuadtreeNode::SharedPtr pParent)
+{
+  std::vector<double> rows = {mfRowBegin, (mfRowBegin + mfRowEnd) / 2, mfRowEnd};
+  std::vector<double> cols = {mfColBegin, (mfColBegin + mfColEnd) / 2, mfColEnd};
+  for (int i = 0; i < 2; ++i)
+  {
+    for (int j = 0; j < 2; ++j)
+    {
+      auto node = std::make_shared<QuadtreeNode>(pParent, rows[i], rows[i + 1], cols[j], cols[j + 1]);
+      mvpChildren.push_back(node);
     }
+  }
 }
 
 /**
@@ -71,19 +78,21 @@ void QuadtreeNode::split(QuadtreeNode::SharedPtr pParent) {
  *      2. 根据这个节点比例，进行初始节点的初始化
  * @param pParent
  */
-void QuadtreeNode::initSplit(SharedPtr pParent) {
-    const double &w = mfColEnd;
-    const double &h = mfRowEnd;
-    const int nIni = round(w / h);
-    const float hX = (double)w / nIni;
-    std::vector<double> cols = {mfColBegin};
-    for (std::size_t idx = 1; idx < nIni; ++idx)
-        cols.push_back(idx * hX);
-    cols.push_back(mfColEnd);
-    for (int i = 0; i < nIni; ++i) {
-        auto node = std::make_shared<QuadtreeNode>(pParent, mfRowBegin, mfRowEnd, cols[i], cols[i + 1]);
-        mvpChildren.push_back(node);
-    }
+void QuadtreeNode::initSplit(SharedPtr pParent)
+{
+  const double &w = mfColEnd;
+  const double &h = mfRowEnd;
+  const int nIni = round(w / h);
+  const float hX = (double)w / nIni;
+  std::vector<double> cols = {mfColBegin};
+  for (std::size_t idx = 1; idx < nIni; ++idx)
+    cols.push_back(idx * hX);
+  cols.push_back(mfColEnd);
+  for (int i = 0; i < nIni; ++i)
+  {
+    auto node = std::make_shared<QuadtreeNode>(pParent, mfRowBegin, mfRowEnd, cols[i], cols[i + 1]);
+    mvpChildren.push_back(node);
+  }
 }
 
 /**
@@ -91,17 +100,20 @@ void QuadtreeNode::initSplit(SharedPtr pParent) {
  *
  * @return std::size_t 最大响应特征点的索引
  */
-std::size_t QuadtreeNode::getFeature() {
-    std::size_t maxResIdx = 0;
-    float maxResponse = 0.0;
-    for (const auto &idx : mvKeyPoints) {
-        auto &kp = mpAllKeyPoints->at(idx);
-        if (kp.response > maxResponse) {
-            maxResIdx = idx;
-            maxResponse = kp.response;
-        }
+std::size_t QuadtreeNode::getFeature()
+{
+  std::size_t maxResIdx = 0;
+  float maxResponse = 0.0;
+  for (const auto &idx : mvKeyPoints)
+  {
+    auto &kp = mpAllKeyPoints->at(idx);
+    if (kp.response > maxResponse)
+    {
+      maxResIdx = idx;
+      maxResponse = kp.response;
     }
-    return std::move(maxResIdx);
+  }
+  return std::move(maxResIdx);
 }
 
 /**
@@ -112,13 +124,14 @@ std::size_t QuadtreeNode::getFeature() {
  * @param needNodes 需要的特征点数目
  */
 Quadtree::Quadtree(int x, int y, const std::vector<cv::KeyPoint> &keypoints, unsigned int needNodes)
-    : mnNeedNodes(needNodes) {
-    // if (keypoints.size() < mnNeedNodes) {
-    //     throw FeatureLessError("特征点数目少于期待特征点数目，无法分裂");
-    // }
-    mpRoot = std::make_shared<QuadtreeNode>(x, y, keypoints);
-    mvNodes.insert(mpRoot);
-    mnNodes = 1;
+    : mnNeedNodes(needNodes)
+{
+  // if (keypoints.size() < mnNeedNodes) {
+  //     throw FeatureLessError("特征点数目少于期待特征点数目，无法分裂");
+  // }
+  mpRoot = std::make_shared<QuadtreeNode>(x, y, keypoints);
+  mvNodes.insert(mpRoot);
+  mnNodes = 1;
 }
 
 /**
@@ -128,30 +141,35 @@ Quadtree::Quadtree(int x, int y, const std::vector<cv::KeyPoint> &keypoints, uns
  *      2. 在每一次迭代过程中，只对map的第一个节点进行分裂，然后将分裂后的节点去掉
  *      3. 重复上述过程，直到节点数目大于等于需要的节点数目
  */
-void Quadtree::split() {
-    std::multimap<std::size_t, QuadtreeNode::SharedPtr, std::greater<std::size_t>> toSplitNodes;
-    mpRoot->initSplit(mpRoot);
-    toSplitNodes.insert(std::make_pair(mpRoot->mvKeyPoints.size(), mpRoot));
-    mnNodes = 1;
-    // 值得注意的是，toSplitNodes这种情况不可能出现
-    while (mnNodes < mnNeedNodes && !toSplitNodes.empty()) {
-        auto toSplitNodeIter = toSplitNodes.begin();
-        auto toSplitNode = toSplitNodeIter->second;
-        if (toSplitNode != mpRoot) {
-            toSplitNode->split(toSplitNode);
-        }
-        toSplitNodes.erase(toSplitNodeIter);
-        mnNodes -= 1;
-        for (auto &childShared : toSplitNode->mvpChildren) {
-            mvNodes.insert(childShared);
-            if (childShared->mvKeyPoints.empty()) {
-                continue;
-            }
-            toSplitNodes.insert(std::make_pair(childShared->mvKeyPoints.size(), childShared));
-            mnNodes += 1;
-        }
+void Quadtree::split()
+{
+  std::multimap<std::size_t, QuadtreeNode::SharedPtr, std::greater<std::size_t>> toSplitNodes;
+  mpRoot->initSplit(mpRoot);
+  toSplitNodes.insert(std::make_pair(mpRoot->mvKeyPoints.size(), mpRoot));
+  mnNodes = 1;
+  // 值得注意的是，toSplitNodes这种情况不可能出现
+  while (mnNodes < mnNeedNodes && !toSplitNodes.empty())
+  {
+    auto toSplitNodeIter = toSplitNodes.begin();
+    auto toSplitNode = toSplitNodeIter->second;
+    if (toSplitNode != mpRoot)
+    {
+      toSplitNode->split(toSplitNode);
     }
-    nodes2kpoints(toSplitNodes);
+    toSplitNodes.erase(toSplitNodeIter);
+    mnNodes -= 1;
+    for (auto &childShared : toSplitNode->mvpChildren)
+    {
+      mvNodes.insert(childShared);
+      if (childShared->mvKeyPoints.empty())
+      {
+        continue;
+      }
+      toSplitNodes.insert(std::make_pair(childShared->mvKeyPoints.size(), childShared));
+      mnNodes += 1;
+    }
+  }
+  nodes2kpoints(toSplitNodes);
 }
 
 /**
@@ -161,14 +179,16 @@ void Quadtree::split() {
  *      2. 只取前mnNeedNodes个节点中的特征点
  * @param toSplitNodes 待分裂的节点，也就是经过分裂留下来的节点
  */
-void Quadtree::nodes2kpoints(WeightNodes &toSplitNodes) {
-    auto iter = toSplitNodes.begin();
-    std::size_t nodeNum = std::min((std::size_t)mnNeedNodes, toSplitNodes.size());
-    for (unsigned i = 0; i < nodeNum; ++i) {
-        auto featIdx = iter->second->getFeature();
-        mvRetKpoints.insert(featIdx);
-        ++iter;
-    }
+void Quadtree::nodes2kpoints(WeightNodes &toSplitNodes)
+{
+  auto iter = toSplitNodes.begin();
+  std::size_t nodeNum = std::min((std::size_t)mnNeedNodes, toSplitNodes.size());
+  for (unsigned i = 0; i < nodeNum; ++i)
+  {
+    auto featIdx = iter->second->getFeature();
+    mvRetKpoints.insert(featIdx);
+    ++iter;
+  }
 }
 
 /**
@@ -182,59 +202,68 @@ void Quadtree::nodes2kpoints(WeightNodes &toSplitNodes) {
  * @param maxThreshold  输入的FAST检测的最大阈值
  * @param minThreshold  输入的FAST检测的最小阈值
  */
-ORBExtractor::ORBExtractor(const cv::Mat &image, int nFeatures, int pyramidLevels, float scaleFactor,
-                           const std::string &bfTemFp, int maxThreshold, int minThreshold)
+ORBExtractor::ORBExtractor(const cv::Mat &image, int nFeatures, int pyramidLevels, float scaleFactor, const std::string &bfTemFp, int maxThreshold,
+                           int minThreshold)
     : mnFeats(std::move(nFeatures))
     , mnMaxThresh(std::move(maxThreshold))
-    , mnMinThresh(std::move(minThreshold)) {
-    initPyramid(image, pyramidLevels, scaleFactor);
-    initBriefTemplate(bfTemFp);
-    initMaxU();
+    , mnMinThresh(std::move(minThreshold))
+{
+  initPyramid(image, pyramidLevels, scaleFactor);
+  initBriefTemplate(bfTemFp);
+  initMaxU();
 }
 
 /// 初始化最大列索引（灰度质心）
-void ORBExtractor::initMaxU() {
-    if (!mbMaxColInit) {
-        mvMaxColIdx.resize(mnCentroidR + 1);
-        int v, v0, vmax = cvFloor(mnCentroidR * sqrt(2.f) / 2 + 1);
-        int vmin = cvCeil(mnCentroidR * sqrt(2.f) / 2);
-        const double hp2 = mnCentroidR * mnCentroidR;
-        for (v = 0; v <= vmax; ++v)
-            mvMaxColIdx[v] = cvRound(sqrt(hp2 - v * v));
-        for (v = mnCentroidR, v0 = 0; v >= vmin; --v) {
-            while (mvMaxColIdx[v0] == mvMaxColIdx[v0 + 1])
-                ++v0;
-            mvMaxColIdx[v] = v0;
-            ++v0;
-        }
-        mbMaxColInit = true;
+void ORBExtractor::initMaxU()
+{
+  if (!mbMaxColInit)
+  {
+    mvMaxColIdx.resize(mnCentroidR + 1);
+    int v, v0, vmax = cvFloor(mnCentroidR * sqrt(2.f) / 2 + 1);
+    int vmin = cvCeil(mnCentroidR * sqrt(2.f) / 2);
+    const double hp2 = mnCentroidR * mnCentroidR;
+    for (v = 0; v <= vmax; ++v)
+      mvMaxColIdx[v] = cvRound(sqrt(hp2 - v * v));
+    for (v = mnCentroidR, v0 = 0; v >= vmin; --v)
+    {
+      while (mvMaxColIdx[v0] == mvMaxColIdx[v0 + 1])
+        ++v0;
+      mvMaxColIdx[v] = v0;
+      ++v0;
     }
+    mbMaxColInit = true;
+  }
 }
 /**
  * @brief 初始化BRIEF模版
  *
  * @param tempFp    BRIEF模版文件的路径
  */
-void ORBExtractor::initBriefTemplate(const std::string &tempFp) {
-    if (!mbTemInit) {
-        std::ifstream ifs(tempFp);
-        if (!ifs.is_open()) {
-            throw FileNotOpenError("BRIEF描述子模版文件打开失败！");
-        }
-        std::string lineStr;
-        bool isHeader = true;
-        while (std::getline(ifs, lineStr)) {
-            if (isHeader) {
-                isHeader = false;
-                continue;
-            }
-            std::istringstream iss(lineStr);
-            cv::Point2f pt1, pt2;
-            iss >> pt1.x >> pt1.y >> pt2.x >> pt2.y;
-            mvBriefTem.push_back(std::make_pair(pt1, pt2));
-        }
-        mbTemInit = true;
+void ORBExtractor::initBriefTemplate(const std::string &tempFp)
+{
+  if (!mbTemInit)
+  {
+    std::ifstream ifs(tempFp);
+    if (!ifs.is_open())
+    {
+      throw FileNotOpenError("BRIEF描述子模版文件打开失败！");
     }
+    std::string lineStr;
+    bool isHeader = true;
+    while (std::getline(ifs, lineStr))
+    {
+      if (isHeader)
+      {
+        isHeader = false;
+        continue;
+      }
+      std::istringstream iss(lineStr);
+      cv::Point2f pt1, pt2;
+      iss >> pt1.x >> pt1.y >> pt2.x >> pt2.y;
+      mvBriefTem.push_back(std::make_pair(pt1, pt2));
+    }
+    mbTemInit = true;
+  }
 }
 
 /**
@@ -246,43 +275,48 @@ void ORBExtractor::initBriefTemplate(const std::string &tempFp) {
  * @param nLevels       输入的金字塔层数
  * @param scaleFactor   输入的金字塔的缩放因子
  */
-void ORBExtractor::initPyramid(const cv::Mat &image, int nLevels, float scaleFactor) {
-    mvPyramids.resize(nLevels);
-    mvBriefMat.resize(nLevels);
+void ORBExtractor::initPyramid(const cv::Mat &image, int nLevels, float scaleFactor)
+{
+  mvPyramids.resize(nLevels);
+  mvBriefMat.resize(nLevels);
 
-    if (!mbScaleInit) {
-        for (int level = 0; level < nLevels; ++level)
-            mvfScaledFactors.push_back(std::pow(scaleFactor, level));
-        mfScaledFactor = scaleFactor;
-        mbScaleInit = true;
-    }
+  if (!mbScaleInit)
+  {
+    for (int level = 0; level < nLevels; ++level)
+      mvfScaledFactors.push_back(std::pow(scaleFactor, level));
+    mfScaledFactor = scaleFactor;
+    mbScaleInit = true;
+  }
 
-    mvnFeatures.resize(nLevels, 0);
-    float scale = 1.0f / mfScaledFactor;
-    int sumFeatures = 0;
-    int nfeats = cvRound(mnFeats * (1 - scale) / (1 - std::pow(scale, nLevels)));
-    for (int level = 0; level < nLevels - 1; ++level) {
-        mvnFeatures[level] = nfeats;
-        sumFeatures += nfeats;
-        nfeats = cvRound(nfeats * scale);
-    }
-    mvnFeatures[nLevels - 1] = std::max(0, mnFeats - sumFeatures);
-    mnLevels = nLevels;
+  mvnFeatures.resize(nLevels, 0);
+  float scale = 1.0f / mfScaledFactor;
+  int sumFeatures = 0;
+  int nfeats = cvRound(mnFeats * (1 - scale) / (1 - std::pow(scale, nLevels)));
+  for (int level = 0; level < nLevels - 1; ++level)
+  {
+    mvnFeatures[level] = nfeats;
+    sumFeatures += nfeats;
+    nfeats = cvRound(nfeats * scale);
+  }
+  mvnFeatures[nLevels - 1] = std::max(0, mnFeats - sumFeatures);
+  mnLevels = nLevels;
 
-    image.copyTo(mvPyramids[0]);
-    int width = image.cols, height = image.rows;
-    for (int i = 1; i < nLevels; ++i) {
-        int levelWidth = cvRound(width / mvfScaledFactors[i]);
-        int levelHeight = cvRound(height / mvfScaledFactors[i]);
-        if (levelWidth < 2 * mnBorderSize || levelHeight < 2 * mnBorderSize) {
-            throw ImageSizeError("金字塔缩放的图像尺寸有问题！");
-            return;
-        }
-        cv::Size sz(levelWidth, levelHeight);
-        cv::resize(image, mvPyramids[i], sz, 0, 0, cv::INTER_LINEAR);
+  image.copyTo(mvPyramids[0]);
+  int width = image.cols, height = image.rows;
+  for (int i = 1; i < nLevels; ++i)
+  {
+    int levelWidth = cvRound(width / mvfScaledFactors[i]);
+    int levelHeight = cvRound(height / mvfScaledFactors[i]);
+    if (levelWidth < 2 * mnBorderSize || levelHeight < 2 * mnBorderSize)
+    {
+      throw ImageSizeError("金字塔缩放的图像尺寸有问题！");
+      return;
     }
-    for (int i = 0; i < nLevels; ++i)
-        cv::GaussianBlur(mvPyramids[i], mvBriefMat[i], cv::Size(7, 7), 2, 2, cv::BORDER_REFLECT_101);
+    cv::Size sz(levelWidth, levelHeight);
+    cv::resize(image, mvPyramids[i], sz, 0, 0, cv::INTER_LINEAR);
+  }
+  for (int i = 0; i < nLevels; ++i)
+    cv::GaussianBlur(mvPyramids[i], mvBriefMat[i], cv::Size(7, 7), 2, 2, cv::BORDER_REFLECT_101);
 }
 
 /**
@@ -294,57 +328,62 @@ void ORBExtractor::initPyramid(const cv::Mat &image, int nLevels, float scaleFac
  * @param nlevel    输入的金字塔层级
  * @param keyPoints 输出的提取到的FAST角点vector
  */
-void ORBExtractor::extractFast(int nlevel, std::vector<cv::KeyPoint> &keyPoints) {
-    cv::Mat image = mvPyramids[nlevel];
-    int nMaxBorderX = image.cols - mnBorderSize + 3;
-    int nMaxBorderY = image.rows - mnBorderSize + 3;
-    int nMinBorderX = mnBorderSize - 3;
-    int nMinBorderY = mnBorderSize - 3;
-    int w = nMaxBorderX - nMinBorderX;
-    int h = nMaxBorderY - nMinBorderY;
-    const int nCols = w / 30;
-    const int nRows = h / 30;
-    const int wCell = ceil(w / nCols);
-    const int hCell = ceil(h / nRows);
+void ORBExtractor::extractFast(int nlevel, std::vector<cv::KeyPoint> &keyPoints)
+{
+  cv::Mat image = mvPyramids[nlevel];
+  int nMaxBorderX = image.cols - mnBorderSize + 3;
+  int nMaxBorderY = image.rows - mnBorderSize + 3;
+  int nMinBorderX = mnBorderSize - 3;
+  int nMinBorderY = mnBorderSize - 3;
+  int w = nMaxBorderX - nMinBorderX;
+  int h = nMaxBorderY - nMinBorderY;
+  const int nCols = w / 30;
+  const int nRows = h / 30;
+  const int wCell = ceil(w / nCols);
+  const int hCell = ceil(h / nRows);
 
-    std::vector<cv::KeyPoint> levelKps;
-    for (std::size_t idx = 0; idx < nRows; ++idx) {
-        int iniY = nMinBorderY + idx * hCell;
-        int maxY = iniY + hCell + 6;
-        if (iniY >= nMaxBorderY - 6)
-            continue;
-        if (maxY > nMaxBorderY)
-            maxY = nMaxBorderY;
+  std::vector<cv::KeyPoint> levelKps;
+  for (std::size_t idx = 0; idx < nRows; ++idx)
+  {
+    int iniY = nMinBorderY + idx * hCell;
+    int maxY = iniY + hCell + 6;
+    if (iniY >= nMaxBorderY - 6)
+      continue;
+    if (maxY > nMaxBorderY)
+      maxY = nMaxBorderY;
 
-        for (std::size_t jdx = 0; jdx < nCols; ++jdx) {
-            int iniX = nMinBorderX + jdx * wCell;
-            int maxX = iniX + wCell + 6;
-            if (iniX >= nMaxBorderX - 6)
-                continue;
-            if (maxX > nMaxBorderX)
-                maxX = nMaxBorderX;
-            cv::Mat pitch = image.rowRange(iniY, maxY).colRange(iniX, maxX);
-            std::vector<cv::KeyPoint> pitchKps;
-            cv::FAST(pitch, pitchKps, mnMaxThresh, true);
-            if (pitchKps.empty())
-                cv::FAST(pitch, pitchKps, mnMinThresh, true);
-            for (auto &kp : pitchKps) {
-                kp.pt.x += jdx * wCell;
-                kp.pt.y += idx * hCell;
-            }
-            std::copy(pitchKps.begin(), pitchKps.end(), std::back_inserter(levelKps));
-        }
+    for (std::size_t jdx = 0; jdx < nCols; ++jdx)
+    {
+      int iniX = nMinBorderX + jdx * wCell;
+      int maxX = iniX + wCell + 6;
+      if (iniX >= nMaxBorderX - 6)
+        continue;
+      if (maxX > nMaxBorderX)
+        maxX = nMaxBorderX;
+      cv::Mat pitch = image.rowRange(iniY, maxY).colRange(iniX, maxX);
+      std::vector<cv::KeyPoint> pitchKps;
+      cv::FAST(pitch, pitchKps, mnMaxThresh, true);
+      if (pitchKps.empty())
+        cv::FAST(pitch, pitchKps, mnMinThresh, true);
+      for (auto &kp : pitchKps)
+      {
+        kp.pt.x += jdx * wCell;
+        kp.pt.y += idx * hCell;
+      }
+      std::copy(pitchKps.begin(), pitchKps.end(), std::back_inserter(levelKps));
     }
-    Quadtree quadTree(w, h, levelKps, mvnFeatures[nlevel]);
-    quadTree.split();
-    auto idxs = quadTree.getFeatIdxs();
-    for (const auto &id : idxs) {
-        auto &kp = levelKps[id];
-        kp.pt.x += nMinBorderX;
-        kp.pt.y += nMinBorderY;
-        kp.octave = nlevel;
-        keyPoints.push_back(kp);
-    }
+  }
+  Quadtree quadTree(w, h, levelKps, mvnFeatures[nlevel]);
+  quadTree.split();
+  auto idxs = quadTree.getFeatIdxs();
+  for (const auto &id : idxs)
+  {
+    auto &kp = levelKps[id];
+    kp.pt.x += nMinBorderX;
+    kp.pt.y += nMinBorderY;
+    kp.octave = nlevel;
+    keyPoints.push_back(kp);
+  }
 }
 
 /**
@@ -355,22 +394,24 @@ void ORBExtractor::extractFast(int nlevel, std::vector<cv::KeyPoint> &keyPoints)
  * @param keyPoints     输入的关键点集合
  * @return 描述子(cv::Mat)
  */
-std::vector<cv::Mat> ORBExtractor::computeBRIEF(std::vector<cv::KeyPoint> &keyPoints) {
-    std::vector<cv::Mat> descriptors;
-    for (int idx = 0; idx < keyPoints.size(); ++idx) {
-        cv::Mat descriptorCV(1, 32, CV_8U);
-        auto &keypoint = keyPoints[idx];
-        auto &sacledFactor = mvfScaledFactors[keypoint.octave];
-        std::vector<uchar> descriptor;
-        double angle = computeBRIEF(keypoint.octave, keypoint.pt, descriptor);
-        keypoint.angle = angle / M_PI * 180;
-        keypoint.pt.x *= sacledFactor;
-        keypoint.pt.y *= sacledFactor;
-        for (int i = 0; i < 32; ++i)
-            descriptorCV.at<uchar>(0, i) = descriptor[i];
-        descriptors.push_back(descriptorCV);
-    }
-    return descriptors;
+std::vector<cv::Mat> ORBExtractor::computeBRIEF(std::vector<cv::KeyPoint> &keyPoints)
+{
+  std::vector<cv::Mat> descriptors;
+  for (int idx = 0; idx < keyPoints.size(); ++idx)
+  {
+    cv::Mat descriptorCV(1, 32, CV_8U);
+    auto &keypoint = keyPoints[idx];
+    auto &sacledFactor = mvfScaledFactors[keypoint.octave];
+    std::vector<uchar> descriptor;
+    double angle = computeBRIEF(keypoint.octave, keypoint.pt, descriptor);
+    keypoint.angle = angle / M_PI * 180;
+    keypoint.pt.x *= sacledFactor;
+    keypoint.pt.y *= sacledFactor;
+    for (int i = 0; i < 32; ++i)
+      descriptorCV.at<uchar>(0, i) = descriptor[i];
+    descriptors.push_back(descriptorCV);
+  }
+  return descriptors;
 }
 
 /**
@@ -383,32 +424,35 @@ std::vector<cv::Mat> ORBExtractor::computeBRIEF(std::vector<cv::KeyPoint> &keyPo
  * @param descriptor    输出的关键点的描述子（32的行向量）
  * @return double       输出的point点对应的角度信息（弧度）
  */
-double ORBExtractor::computeBRIEF(const int &nLevel, const cv::Point2f &point, std::vector<uchar> &descriptor) {
-    descriptor.clear();
-    descriptor.reserve(32);
-    const cv::Mat &image = mvPyramids[nLevel];
-    const cv::Mat &workMat = mvBriefMat[nLevel];
-    double theta = getGrayCentroid(image, point);
-    double cosVal = std::cos(theta);
-    double sinVal = std::sin(theta);
+double ORBExtractor::computeBRIEF(const int &nLevel, const cv::Point2f &point, std::vector<uchar> &descriptor)
+{
+  descriptor.clear();
+  descriptor.reserve(32);
+  const cv::Mat &image = mvPyramids[nLevel];
+  const cv::Mat &workMat = mvBriefMat[nLevel];
+  double theta = getGrayCentroid(image, point);
+  double cosVal = std::cos(theta);
+  double sinVal = std::sin(theta);
 
-    uchar value = 0;
-    int bias = 0;
-    for (auto &pointPair : mvBriefTem) {
-        cv::Point2f p1 = rotateTemplate(pointPair.first, sinVal, cosVal);
-        cv::Point2f p2 = rotateTemplate(pointPair.second, sinVal, cosVal);
-        uchar p1Val = workMat.at<uchar>(cvRound(point.y + p1.y), cvRound(point.x + p1.x));
-        uchar p2Val = workMat.at<uchar>(cvRound(point.y + p2.y), cvRound(point.x + p2.x));
-        value |= (p1Val < p2Val) << bias;
-        if (bias == 7) {
-            descriptor.push_back(value);
-            bias = 0;
-            value = 0;
-            continue;
-        }
-        ++bias;
+  uchar value = 0;
+  int bias = 0;
+  for (auto &pointPair : mvBriefTem)
+  {
+    cv::Point2f p1 = rotateTemplate(pointPair.first, sinVal, cosVal);
+    cv::Point2f p2 = rotateTemplate(pointPair.second, sinVal, cosVal);
+    uchar p1Val = workMat.at<uchar>(cvRound(point.y + p1.y), cvRound(point.x + p1.x));
+    uchar p2Val = workMat.at<uchar>(cvRound(point.y + p2.y), cvRound(point.x + p2.x));
+    value |= (p1Val < p2Val) << bias;
+    if (bias == 7)
+    {
+      descriptor.push_back(value);
+      bias = 0;
+      value = 0;
+      continue;
     }
-    return theta;
+    ++bias;
+  }
+  return theta;
 }
 
 /**
@@ -418,25 +462,28 @@ double ORBExtractor::computeBRIEF(const int &nLevel, const cv::Point2f &point, s
  * @param point     输入的点信息
  * @return double   输出的灰度质心法计算的角度值
  */
-double ORBExtractor::getGrayCentroid(const cv::Mat &image, const cv::Point2f &point) {
-    int x = cvRound(point.x);
-    int y = cvRound(point.y);
-    int m10 = 0, m01 = 0;
-    for (int dx = -mnCentroidR; dx <= mnCentroidR; ++dx)
-        m10 += dx * image.at<uchar>(y, x + dx);
+double ORBExtractor::getGrayCentroid(const cv::Mat &image, const cv::Point2f &point)
+{
+  int x = cvRound(point.x);
+  int y = cvRound(point.y);
+  int m10 = 0, m01 = 0;
+  for (int dx = -mnCentroidR; dx <= mnCentroidR; ++dx)
+    m10 += dx * image.at<uchar>(y, x + dx);
 
-    for (int dy = 1; dy <= mnCentroidR; ++dy) {
-        int vSum = 0;
-        int d = mvMaxColIdx[dy];
-        for (int dx = -d; dx <= d; ++dx) {
-            uchar upValue = image.at<uchar>(y + dy, x + dx);
-            uchar downValue = image.at<uchar>(y - dy, x + dx);
-            m10 += dx * (upValue + downValue);
-            vSum += (upValue - downValue);
-        }
-        m01 += vSum * dy;
+  for (int dy = 1; dy <= mnCentroidR; ++dy)
+  {
+    int vSum = 0;
+    int d = mvMaxColIdx[dy];
+    for (int dx = -d; dx <= d; ++dx)
+    {
+      uchar upValue = image.at<uchar>(y + dy, x + dx);
+      uchar downValue = image.at<uchar>(y - dy, x + dx);
+      m10 += dx * (upValue + downValue);
+      vSum += (upValue - downValue);
     }
-    return std::atan2((double)m01, (double)m10);
+    m01 += vSum * dy;
+  }
+  return std::atan2((double)m01, (double)m10);
 }
 
 /**
@@ -449,13 +496,15 @@ double ORBExtractor::getGrayCentroid(const cv::Mat &image, const cv::Point2f &po
  * @param keyPoints     输出的特征点
  * @param descriptors   输出的描述子（cv::Mat）
  */
-void ORBExtractor::extract(std::vector<cv::KeyPoint> &keyPoints, std::vector<cv::Mat> &descriptors) {
-    for (int level = 0; level < mnLevels; ++level) {
-        std::vector<cv::KeyPoint> levelKps;
-        extractFast(level, levelKps);
-        std::copy(levelKps.begin(), levelKps.end(), std::back_inserter(keyPoints));
-    }
-    descriptors = computeBRIEF(keyPoints);
+void ORBExtractor::extract(std::vector<cv::KeyPoint> &keyPoints, std::vector<cv::Mat> &descriptors)
+{
+  for (int level = 0; level < mnLevels; ++level)
+  {
+    std::vector<cv::KeyPoint> levelKps;
+    extractFast(level, levelKps);
+    std::copy(levelKps.begin(), levelKps.end(), std::back_inserter(keyPoints));
+  }
+  descriptors = computeBRIEF(keyPoints);
 }
 
 /// ORBExtractor的静态变量类外初始化
@@ -482,11 +531,12 @@ float ORBExtractor::mfScaledFactor;
  * @param cosValue      旋转角度cos值
  * @return cv::Point2f  返回旋转后的点
  */
-cv::Point2f rotateTemplate(const cv::Point2f &toRotPoint, const double &sinValue, const double &cosValue) {
-    cv::Point2f point;
-    point.x = toRotPoint.x * cosValue - toRotPoint.y * sinValue;
-    point.y = toRotPoint.x * sinValue + toRotPoint.y * cosValue;
-    return point;
+cv::Point2f rotateTemplate(const cv::Point2f &toRotPoint, const double &sinValue, const double &cosValue)
+{
+  cv::Point2f point;
+  point.x = toRotPoint.x * cosValue - toRotPoint.y * sinValue;
+  point.y = toRotPoint.x * sinValue + toRotPoint.y * cosValue;
+  return point;
 }
 
 } // namespace ORB_SLAM2_ROS2
