@@ -1,7 +1,5 @@
 #include <iostream>
 
-#include <gflags/gflags.h>
-
 #include "ORB_SLAM2/Camera.h"
 #include "ORB_SLAM2/KeyFrame.h"
 #include "ORB_SLAM2/KeyFrameDB.h"
@@ -13,13 +11,14 @@
 #include "ORB_SLAM2/Tracking.h"
 #include "ORB_SLAM2/Viewer.h"
 
-namespace ORB_SLAM2_ROS2 {
+namespace ORB_SLAM2_ROS2
+{
 
 using namespace std::chrono_literals;
 
-Tracking::Tracking(Map::SharedPtr pMap, KFrameDBPtr pKfDB, VocPtr pVoc, const std::string &sBriefTemplate,
-                   int nFeatures, int nInitFeatures, int nInitFAST, int nMinFAST, int nDepthTh, int nMaxFrame,
-                   int nMinFrame, int nColorType, bool bOnlyTracking, float fDepthScale, int nLevel, float fScaleFactor)
+Tracking::Tracking(Map::SharedPtr pMap, KFrameDBPtr pKfDB, VocPtr pVoc, const std::string &sBriefTemplate, int nFeatures, int nInitFeatures, int nInitFAST,
+                   int nMinFAST, int nDepthTh, int nMaxFrame, int nMinFrame, int nColorType, bool bOnlyTracking, float fDepthScale, int nLevel,
+                   float fScaleFactor)
     : mpMap(pMap)
     , mpKfDB(pKfDB)
     , mpVoc(pVoc)
@@ -39,7 +38,9 @@ Tracking::Tracking(Map::SharedPtr pMap, KFrameDBPtr pKfDB, VocPtr pVoc, const st
     , mStatus(TrackingState::NOT_IMAGE_YET)
     , mfdScale(fDepthScale)
     , mnLevels(nLevel)
-    , mfScaleFactor(fScaleFactor) {}
+    , mfScaleFactor(fScaleFactor)
+{
+}
 
 /**
  * @brief 处理普通帧
@@ -48,141 +49,164 @@ Tracking::Tracking(Map::SharedPtr pMap, KFrameDBPtr pKfDB, VocPtr pVoc, const st
  * @param rightImg  右图像
  * @return cv::Mat 输出的位姿矩阵
  */
-cv::Mat Tracking::grabFrame(cv::Mat leftImg, cv::Mat rightImg) {
-    mleftImg = leftImg;
-    if (mnColorType == 1) {
-        /// 对应的RGB图像
-        cv::cvtColor(leftImg, leftImg, cv::COLOR_RGB2GRAY);
-        if (Camera::mType == CameraType::Stereo)
-            cv::cvtColor(rightImg, rightImg, cv::COLOR_RGB2GRAY);
-    } else if (mnColorType == 2) {
-        /// 对应的BGR图像
-        cv::cvtColor(leftImg, leftImg, cv::COLOR_BGR2GRAY);
-        if (Camera::mType == CameraType::Stereo)
-            cv::cvtColor(rightImg, rightImg, cv::COLOR_BGR2GRAY);
-    }
-    // switch (mStatus) {
-    // case TrackingState::NOT_IMAGE_YET:
-    //     mpCurrFrame =
-    //         Frame::createStereo(leftImg, rightImg, mnInitFeatures, msBriefTemFp, mnMaxThresh, mnMinThresh, mpVoc);
-    //     break;
-    // case TrackingState::NOT_INITING:
-    //     mpCurrFrame =
-    //         Frame::createStereo(leftImg, rightImg, mnInitFeatures, msBriefTemFp, mnMaxThresh, mnMinThresh, mpVoc);
-    //     break;
-    // default:
-    //     mpCurrFrame = Frame::createStereo(leftImg, rightImg, mnFeatures, msBriefTemFp, mnMaxThresh, mnMinThresh,
-    //     mpVoc);
-    // }
-    switch (Camera::mType) {
-    case CameraType::Stereo:
-        mpCurrFrame = Frame::createStereo(leftImg, rightImg, mnFeatures, msBriefTemFp, mnMaxThresh, mnMinThresh, mpVoc,
-                                          mnLevels, mfScaleFactor);
-        break;
-    case CameraType::RGBD:
-        mpCurrFrame = Frame::createRGBD(leftImg, rightImg, mnFeatures, msBriefTemFp, mnMaxThresh, mnMinThresh, mpVoc,
-                                        mfdScale, mnLevels, mfScaleFactor);
-        break;
-    default:
-        throw std::runtime_error("此ORB_SLAM2只能在双目或者RGBD图像中使用");
-        break;
-    }
+cv::Mat Tracking::grabFrame(cv::Mat leftImg, cv::Mat rightImg)
+{
+  mleftImg = leftImg;
+  if (mnColorType == 1)
+  {
+    /// 对应的RGB图像
+    cv::cvtColor(leftImg, leftImg, cv::COLOR_RGB2GRAY);
+    if (Camera::mType == CameraType::Stereo)
+      cv::cvtColor(rightImg, rightImg, cv::COLOR_RGB2GRAY);
+  }
+  else if (mnColorType == 2)
+  {
+    /// 对应的BGR图像
+    cv::cvtColor(leftImg, leftImg, cv::COLOR_BGR2GRAY);
+    if (Camera::mType == CameraType::Stereo)
+      cv::cvtColor(rightImg, rightImg, cv::COLOR_BGR2GRAY);
+  }
+  // switch (mStatus) {
+  // case TrackingState::NOT_IMAGE_YET:
+  //     mpCurrFrame =
+  //         Frame::createStereo(leftImg, rightImg, mnInitFeatures, msBriefTemFp, mnMaxThresh, mnMinThresh, mpVoc);
+  //     break;
+  // case TrackingState::NOT_INITING:
+  //     mpCurrFrame =
+  //         Frame::createStereo(leftImg, rightImg, mnInitFeatures, msBriefTemFp, mnMaxThresh, mnMinThresh, mpVoc);
+  //     break;
+  // default:
+  //     mpCurrFrame = Frame::createStereo(leftImg, rightImg, mnFeatures, msBriefTemFp, mnMaxThresh, mnMinThresh,
+  //     mpVoc);
+  // }
+  switch (Camera::mType)
+  {
+  case CameraType::Stereo:
+    mpCurrFrame = Frame::createStereo(leftImg, rightImg, mnFeatures, msBriefTemFp, mnMaxThresh, mnMinThresh, mpVoc, mnLevels, mfScaleFactor);
+    break;
+  case CameraType::RGBD:
+    mpCurrFrame = Frame::createRGBD(leftImg, rightImg, mnFeatures, msBriefTemFp, mnMaxThresh, mnMinThresh, mpVoc, mfdScale, mnLevels, mfScaleFactor);
+    break;
+  default:
+    throw std::runtime_error("此ORB_SLAM2只能在双目或者RGBD图像中使用");
+    break;
+  }
 
-    bool bOK = false;
-    if (!mbOnlyTracking) {
-        /// 初始化阶段
-        if (mStatus == TrackingState::NOT_IMAGE_YET || mStatus == TrackingState::NOT_INITING) {
-            mStatus = TrackingState::NOT_INITING;
-            auto pose = cv::Mat::eye(4, 4, CV_32F);
-            mpCurrFrame->setPose(pose);
-            if (mpCurrFrame->getN() >= 500) {
-                initForStereo();
-                mpLastFrame = mpCurrFrame;
-                if (mpViewer)
-                    mpViewer->setCurrFrame(mleftImg, mpCurrFrame, mStatus);
-            }
-            return mpCurrFrame->getPose();
-        }
-    } else {
-        if (mStatus == TrackingState::NOT_IMAGE_YET)
-            mStatus = TrackingState::LOST;
-    }
-    if (mStatus == TrackingState::OK) {
-        mpCurrFrame->setPose(mpLastFrame->getPose());
-        if (mbOnlyTracking || mVelocity.empty() || mpCurrFrame->getID() < mnLastRelocId + 2) {
-            bOK = trackReference();
-        } else {
-            bOK = trackMotionModel();
-            if (!bOK)
-                bOK = trackReference();
-        }
-    } else
-        bOK = trackReLocalize();
-    if (bOK)
-        bOK = trackLocalMap();
-
-    if (bOK) {
-        mStatus = TrackingState::OK;
-        updateVelocity();
-        updateTlr();
-        if (mpLocalMapper && !mbOnlyTracking)
-            insertKeyFrame();
-
+  bool bOK = false;
+  if (!mbOnlyTracking)
+  {
+    /// 初始化阶段
+    if (mStatus == TrackingState::NOT_IMAGE_YET || mStatus == TrackingState::NOT_INITING)
+    {
+      mStatus = TrackingState::NOT_INITING;
+      auto pose = cv::Mat::eye(4, 4, CV_32F);
+      mpCurrFrame->setPose(pose);
+      if (mpCurrFrame->getN() >= 500)
+      {
+        initForStereo();
         mpLastFrame = mpCurrFrame;
         if (mpViewer)
-            mpViewer->setCurrFrame(mleftImg, mpCurrFrame, mStatus);
-        if (mbOnlyTracking && mpRefKf && !mpRefKf->isBad()) {
-            mpRefKf->setNotErased(false);
-            mpRefKf = mpMap->getTrackingRef(mpCurrFrame, mpRefKf->getID());
-            mpRefKf->setNotErased(true);
-        }
-        return mpCurrFrame->getPose();
-    } else {
-        mStatus = TrackingState::LOST;
-        if (mpViewer)
-            mpViewer->setCurrFrame(mleftImg, mpCurrFrame, mStatus);
-        return cv::Mat();
+          mpViewer->setCurrFrame(mleftImg, mpCurrFrame, mStatus);
+      }
+      return mpCurrFrame->getPose();
     }
+  }
+  else
+  {
+    if (mStatus == TrackingState::NOT_IMAGE_YET)
+      mStatus = TrackingState::LOST;
+  }
+  if (mStatus == TrackingState::OK)
+  {
+    mpCurrFrame->setPose(mpLastFrame->getPose());
+    if (mbOnlyTracking || mVelocity.empty() || mpCurrFrame->getID() < mnLastRelocId + 2)
+    {
+      bOK = trackReference();
+    }
+    else
+    {
+      bOK = trackMotionModel();
+      if (!bOK)
+        bOK = trackReference();
+    }
+  }
+  else
+    bOK = trackReLocalize();
+  if (bOK)
+    bOK = trackLocalMap();
+
+  if (bOK)
+  {
+    mStatus = TrackingState::OK;
+    updateVelocity();
+    updateTlr();
+    if (mpLocalMapper && !mbOnlyTracking)
+      insertKeyFrame();
+
+    mpLastFrame = mpCurrFrame;
+    if (mpViewer)
+      mpViewer->setCurrFrame(mleftImg, mpCurrFrame, mStatus);
+    if (mbOnlyTracking && mpRefKf && !mpRefKf->isBad())
+    {
+      mpRefKf->setNotErased(false);
+      mpRefKf = mpMap->getTrackingRef(mpCurrFrame, mpRefKf->getID());
+      mpRefKf->setNotErased(true);
+    }
+    return mpCurrFrame->getPose();
+  }
+  else
+  {
+    mStatus = TrackingState::LOST;
+    if (mpViewer)
+      mpViewer->setCurrFrame(mleftImg, mpCurrFrame, mStatus);
+    return cv::Mat();
+  }
 }
 
 /// 向局部建图线程中插入关键帧
-void Tracking::insertKeyFrame() {
-    if (!needNewKeyFrame())
-        return;
-    auto kf = updateCurrFrame();
-    kf->setNotErased(true);
-    mpRefKf->setNotErased(false);
-    mpRefKf = kf;
-    mpLocalMapper->insertKeyFrame(kf);
-    mnLastInsertId = mpCurrFrame->getID();
+void Tracking::insertKeyFrame()
+{
+  if (!needNewKeyFrame())
+    return;
+  auto kf = updateCurrFrame();
+  kf->setNotErased(true);
+  mpRefKf->setNotErased(false);
+  mpRefKf = kf;
+  mpLocalMapper->insertKeyFrame(kf);
+  mnLastInsertId = mpCurrFrame->getID();
 }
 
 /// 将当前帧升级为关键帧
-KeyFrame::SharedPtr Tracking::updateCurrFrame() {
-    std::vector<MapPoint::SharedPtr> mps;
-    int nCreated = mpCurrFrame->unProject(mps);
-    return KeyFrame::create(*mpCurrFrame);
+KeyFrame::SharedPtr Tracking::updateCurrFrame()
+{
+  std::vector<MapPoint::SharedPtr> mps;
+  int nCreated = mpCurrFrame->unProject(mps);
+  return KeyFrame::create(*mpCurrFrame);
 }
 
 /// 更新速度Tcl
-void Tracking::updateVelocity() {
-    if (!mpLastFrame) {
-        mVelocity = cv::Mat();
-        return;
-    }
-    cv::Mat Tcw = mpCurrFrame->getPose();
-    cv::Mat Twl = mpLastFrame->getPoseInv();
-    mVelocity = Tcw * Twl;
+void Tracking::updateVelocity()
+{
+  if (!mpLastFrame)
+  {
+    mVelocity = cv::Mat();
+    return;
+  }
+  cv::Mat Tcw = mpCurrFrame->getPose();
+  cv::Mat Twl = mpLastFrame->getPoseInv();
+  mVelocity = Tcw * Twl;
 }
 
 /// 更新mTlr（上一帧到参考关键帧的位姿）
-void Tracking::updateTlr() {
-    /// 因为只是用到参考关键帧的速度，因此这里没什么问题
-    KeyFrame::SharedPtr refKF = mpCurrFrame->getRefKF();
-    if (!refKF) {
-        refKF = mpRefKf;
-    }
-    mTlr = mpCurrFrame->getPose() * refKF->getPoseInv();
+void Tracking::updateTlr()
+{
+  /// 因为只是用到参考关键帧的速度，因此这里没什么问题
+  KeyFrame::SharedPtr refKF = mpCurrFrame->getRefKF();
+  if (!refKF)
+  {
+    refKF = mpRefKf;
+  }
+  mTlr = mpCurrFrame->getPose() * refKF->getPoseInv();
 }
 
 /**
@@ -190,31 +214,39 @@ void Tracking::updateTlr() {
  * 插入到局部地图中的关键帧的mbIsLocalKf会被标注为true
  * 同时匹配相同关键的同时插入（判断是否可用）
  */
-void Tracking::insertLocalKFrame(KeyFrame::SharedPtr pKf) {
-    if (pKf && !pKf->isBad()) {
-        auto pParent = pKf->getParent().lock();
-        auto vpChildren = pKf->getChildren();
-        if (pParent && !pParent->isBad()) {
-            if (!pParent->isLocalKf()) {
-                pParent->setLocalKf(true);
-                mvpLocalKfs.push_back(pParent);
-            }
-        }
-        for (auto &childWeak : vpChildren) {
-            auto child = childWeak.lock();
-            if (child && !child->isBad()) {
-                if (!child->isLocalKf()) {
-                    child->setLocalKf(true);
-                    mvpLocalKfs.push_back(child);
-                }
-            }
-        }
-        if (pKf->isLocalKf()) {
-            return;
-        }
-        pKf->setLocalKf(true);
-        mvpLocalKfs.push_back(pKf);
+void Tracking::insertLocalKFrame(KeyFrame::SharedPtr pKf)
+{
+  if (pKf && !pKf->isBad())
+  {
+    auto pParent = pKf->getParent().lock();
+    auto vpChildren = pKf->getChildren();
+    if (pParent && !pParent->isBad())
+    {
+      if (!pParent->isLocalKf())
+      {
+        pParent->setLocalKf(true);
+        mvpLocalKfs.push_back(pParent);
+      }
     }
+    for (auto &childWeak : vpChildren)
+    {
+      auto child = childWeak.lock();
+      if (child && !child->isBad())
+      {
+        if (!child->isLocalKf())
+        {
+          child->setLocalKf(true);
+          mvpLocalKfs.push_back(child);
+        }
+      }
+    }
+    if (pKf->isLocalKf())
+    {
+      return;
+    }
+    pKf->setLocalKf(true);
+    mvpLocalKfs.push_back(pKf);
+  }
 }
 
 /**
@@ -222,13 +254,16 @@ void Tracking::insertLocalKFrame(KeyFrame::SharedPtr pKf) {
  * 插入到局部地图中的地图点的mbIsLocalMp会被标注为true
  * 同时避免重复相同地图点的插入（判断是否可用）
  */
-void Tracking::insertLocalMPoint(MapPoint::SharedPtr pMp) {
-    if (pMp && !pMp->isBad()) {
-        if (!pMp->isLocalMp()) {
-            pMp->setLocalMp(true);
-            mvpLocalMps.push_back(pMp);
-        }
+void Tracking::insertLocalMPoint(MapPoint::SharedPtr pMp)
+{
+  if (pMp && !pMp->isBad())
+  {
+    if (!pMp->isLocalMp())
+    {
+      pMp->setLocalMp(true);
+      mvpLocalMps.push_back(pMp);
     }
+  }
 }
 
 /**
@@ -239,38 +274,43 @@ void Tracking::insertLocalMPoint(MapPoint::SharedPtr pMp) {
  *      3. 上述关键帧的父亲和儿子关键帧
  *      4. 在插入关键帧之前，需要将之前的局部地图关键帧释放，并将mbIsLocalKf置为false
  */
-void Tracking::buildLocalKfs() {
-    for (auto &kf : mvpLocalKfs)
-        kf->setLocalKf(false);
-    mvpLocalKfs.clear();
-    auto firstKfs = mpCurrFrame->getConnectedKfs(0);
-    for (auto &kf : firstKfs) {
-        insertLocalKFrame(kf);
-        auto secondKfs = kf->getConnectedKfs(0);
-        for (auto &kf2 : secondKfs) {
-            insertLocalKFrame(kf2);
-        }
+void Tracking::buildLocalKfs()
+{
+  for (auto &kf : mvpLocalKfs)
+    kf->setLocalKf(false);
+  mvpLocalKfs.clear();
+  auto firstKfs = mpCurrFrame->getConnectedKfs(0);
+  for (auto &kf : firstKfs)
+  {
+    insertLocalKFrame(kf);
+    auto secondKfs = kf->getConnectedKfs(0);
+    for (auto &kf2 : secondKfs)
+    {
+      insertLocalKFrame(kf2);
     }
+  }
 }
 
 /**
  * @brief 构建局部地图中的地图点
  * 注意，在使用这个api之前，保证局部地图中的关键帧已经构建完毕
  */
-void Tracking::buildLocalMps() {
-    {
-        std::unique_lock<std::mutex> lock(mMutexLMps);
-        for (auto &pMp : mvpLocalMps)
-            pMp->setLocalMp(false);
-        mvpLocalMps.clear();
+void Tracking::buildLocalMps()
+{
+  {
+    std::unique_lock<std::mutex> lock(mMutexLMps);
+    for (auto &pMp : mvpLocalMps)
+      pMp->setLocalMp(false);
+    mvpLocalMps.clear();
 
-        for (auto &kf : mvpLocalKfs) {
-            auto vpMps = kf->getMapPoints();
-            for (auto &pMp : vpMps)
-                insertLocalMPoint(pMp);
-        }
+    for (auto &kf : mvpLocalKfs)
+    {
+      auto vpMps = kf->getMapPoints();
+      for (auto &pMp : vpMps)
+        insertLocalMPoint(pMp);
     }
-    setUpdate(true);
+  }
+  setUpdate(true);
 }
 
 /**
@@ -279,31 +319,34 @@ void Tracking::buildLocalMps() {
  *      1. 构建局部地图中的关键帧
  *      2. 构建局部地图中的地图点
  */
-void Tracking::buildLocalMap() {
-    buildLocalKfs();
-    buildLocalMps();
+void Tracking::buildLocalMap()
+{
+  buildLocalKfs();
+  buildLocalMps();
 }
 
 /**
  * @brief 双目相机的初始化
  *
  */
-void Tracking::initForStereo() {
-    std::vector<MapPoint::SharedPtr> mapPoints;
-    mpCurrFrame->unProject(mapPoints);
-    auto kfInit = KeyFrame::create(*mpCurrFrame);
+void Tracking::initForStereo()
+{
+  std::vector<MapPoint::SharedPtr> mapPoints;
+  mpCurrFrame->unProject(mapPoints);
+  auto kfInit = KeyFrame::create(*mpCurrFrame);
 
-    mpMap->insertKeyFrame(kfInit, mpMap);
-    for (std::size_t idx = 0; idx < mapPoints.size(); ++idx) {
-        auto &pMp = mapPoints[idx];
-        if (!pMp)
-            continue;
-        pMp->addAttriInit(kfInit, idx);
-        mpMap->insertMapPoint(pMp, mpMap);
-    }
-    mStatus = TrackingState::OK;
-    mpRefKf = kfInit;
-    mpLocalMapper->insertKeyFrame(kfInit);
+  mpMap->insertKeyFrame(kfInit, mpMap);
+  for (std::size_t idx = 0; idx < mapPoints.size(); ++idx)
+  {
+    auto &pMp = mapPoints[idx];
+    if (!pMp)
+      continue;
+    pMp->addAttriInit(kfInit, idx);
+    mpMap->insertMapPoint(pMp, mpMap);
+  }
+  mStatus = TrackingState::OK;
+  mpRefKf = kfInit;
+  mpLocalMapper->insertKeyFrame(kfInit);
 }
 
 /**
@@ -314,15 +357,17 @@ void Tracking::initForStereo() {
  * @return true     跟踪参考关键帧失败
  * @return false    跟踪参考关键帧成功
  */
-bool Tracking::trackReference() {
-    ORBMatcher matcher(0.7, true);
-    std::vector<cv::DMatch> matches;
-    int nMatches = matcher.searchByBow(mpCurrFrame, mpRefKf, matches);
-    if (nMatches < 10) {
-        return false;
-    }
-    int nInliers = Optimizer::OptimizePoseOnly(mpCurrFrame);
-    return nInliers >= 10;
+bool Tracking::trackReference()
+{
+  ORBMatcher matcher(0.7, true);
+  std::vector<cv::DMatch> matches;
+  int nMatches = matcher.searchByBow(mpCurrFrame, mpRefKf, matches);
+  if (nMatches < 10)
+  {
+    return false;
+  }
+  int nInliers = Optimizer::OptimizePoseOnly(mpCurrFrame);
+  return nInliers >= 10;
 }
 
 /**
@@ -333,27 +378,31 @@ bool Tracking::trackReference() {
  * @return true     跟踪成功
  * @return false    跟踪失败
  */
-bool Tracking::trackMotionModel() {
-    std::vector<cv::DMatch> matches;
-    processLastFrame();
-    mpCurrFrame->setPose(mVelocity * mpLastFrame->getPose());
-    ORBMatcher matcher(0.9, true);
-    int nMatches = matcher.searchByProjection(mpCurrFrame, mpLastFrame, matches, 15);
-    if (nMatches < 20) {
-        nMatches += matcher.searchByProjection(mpCurrFrame, mpLastFrame, matches, 30);
-    }
-    if (nMatches < 20) {
-        return false;
-    }
-    Optimizer::OptimizePoseOnly(mpCurrFrame);
-    int inLiers = 0;
-    auto vMps = mpCurrFrame->getMapPoints();
-    for (std::size_t idx = 0; idx < vMps.size(); ++idx) {
-        auto &pMp = vMps[idx];
-        if (pMp && pMp->isInMap())
-            ++inLiers;
-    }
-    return inLiers >= 10;
+bool Tracking::trackMotionModel()
+{
+  std::vector<cv::DMatch> matches;
+  processLastFrame();
+  mpCurrFrame->setPose(mVelocity * mpLastFrame->getPose());
+  ORBMatcher matcher(0.9, true);
+  int nMatches = matcher.searchByProjection(mpCurrFrame, mpLastFrame, matches, 15);
+  if (nMatches < 20)
+  {
+    nMatches += matcher.searchByProjection(mpCurrFrame, mpLastFrame, matches, 30);
+  }
+  if (nMatches < 20)
+  {
+    return false;
+  }
+  Optimizer::OptimizePoseOnly(mpCurrFrame);
+  int inLiers = 0;
+  auto vMps = mpCurrFrame->getMapPoints();
+  for (std::size_t idx = 0; idx < vMps.size(); ++idx)
+  {
+    auto &pMp = vMps[idx];
+    if (pMp && pMp->isInMap())
+      ++inLiers;
+  }
+  return inLiers >= 10;
 }
 
 /**
@@ -364,12 +413,13 @@ bool Tracking::trackMotionModel() {
  * @return true     找到初步的候选关键帧
  * @return false    没找到初步的候选关键帧
  */
-bool Tracking::findInitialKF(std::vector<KeyFrame::SharedPtr> &vpCandidateKFs, int &candidateNum) {
-    mpKfDB->findRelocKfs(mpCurrFrame, vpCandidateKFs);
-    candidateNum = vpCandidateKFs.size();
-    if (candidateNum == 0)
-        return false;
-    return true;
+bool Tracking::findInitialKF(std::vector<KeyFrame::SharedPtr> &vpCandidateKFs, int &candidateNum)
+{
+  mpKfDB->findRelocKfs(mpCurrFrame, vpCandidateKFs);
+  candidateNum = vpCandidateKFs.size();
+  if (candidateNum == 0)
+    return false;
+  return true;
 }
 
 /**
@@ -377,10 +427,11 @@ bool Tracking::findInitialKF(std::vector<KeyFrame::SharedPtr> &vpCandidateKFs, i
  *
  * @param nCandidate 输入的候选关键帧数量
  */
-RelocBowParam::RelocBowParam(int nCandidate) {
-    mvpSolvers.resize(nCandidate, nullptr);
-    mvAllMatches.resize(nCandidate);
-    mvPnPId2MatchID.resize(nCandidate);
+RelocBowParam::RelocBowParam(int nCandidate)
+{
+  mvpSolvers.resize(nCandidate, nullptr);
+  mvAllMatches.resize(nCandidate);
+  mvPnPId2MatchID.resize(nCandidate);
 }
 
 /**
@@ -393,44 +444,49 @@ RelocBowParam::RelocBowParam(int nCandidate) {
  * @return int 输出的合格关键帧数量
  */
 int Tracking::filterKFByBow(RelocBowParam &relocBowParam, std::vector<bool> &vbDiscard, const int &candidateNum,
-                            std::vector<KeyFrame::SharedPtr> &vpCandidateKFs) {
-    int nCandidates = 0;
-    for (std::size_t idx = 0; idx < candidateNum; ++idx) {
-        KeyFrame::SharedPtr pkf = vpCandidateKFs[idx];
-        if (!pkf || pkf->isBad()) {
-            vbDiscard[idx] = true;
-            continue;
-        }
-        ORBMatcher matcher(0.75, true);
-        std::vector<cv::DMatch> matches;
-        mpCurrFrame->setMapPointsNull();
-        int nMatches = matcher.searchByBow(mpCurrFrame, pkf, matches);
-        relocBowParam.mvAllMatches[idx] = matches;
-        if (nMatches < 10)
-            vbDiscard[idx] = true;
-        else {
-            std::vector<cv::Mat> mapPoints;
-            std::vector<cv::KeyPoint> ORBPoints;
-            const auto &allORBPoints = mpCurrFrame->getLeftKeyPoints();
-            const auto &allMapPoints = pkf->getMapPoints();
-
-            std::size_t pnpIdx = 0;
-            for (std::size_t jdx = 0; jdx < matches.size(); ++jdx) {
-                const auto &match = matches[jdx];
-                const auto &pMp = allMapPoints[match.trainIdx];
-                const auto &ORBPoint = allORBPoints[match.queryIdx];
-                if (!pMp || pMp->isBad())
-                    continue;
-                ORBPoints.push_back(ORBPoint);
-                mapPoints.push_back(pMp->getPos().clone());
-                relocBowParam.mvPnPId2MatchID[idx][pnpIdx] = jdx;
-                ++pnpIdx;
-            }
-            relocBowParam.mvpSolvers[idx] = PnPSolver::create(mapPoints, ORBPoints);
-            ++nCandidates;
-        }
+                            std::vector<KeyFrame::SharedPtr> &vpCandidateKFs)
+{
+  int nCandidates = 0;
+  for (std::size_t idx = 0; idx < candidateNum; ++idx)
+  {
+    KeyFrame::SharedPtr pkf = vpCandidateKFs[idx];
+    if (!pkf || pkf->isBad())
+    {
+      vbDiscard[idx] = true;
+      continue;
     }
-    return nCandidates;
+    ORBMatcher matcher(0.75, true);
+    std::vector<cv::DMatch> matches;
+    mpCurrFrame->setMapPointsNull();
+    int nMatches = matcher.searchByBow(mpCurrFrame, pkf, matches);
+    relocBowParam.mvAllMatches[idx] = matches;
+    if (nMatches < 10)
+      vbDiscard[idx] = true;
+    else
+    {
+      std::vector<cv::Mat> mapPoints;
+      std::vector<cv::KeyPoint> ORBPoints;
+      const auto &allORBPoints = mpCurrFrame->getLeftKeyPoints();
+      const auto &allMapPoints = pkf->getMapPoints();
+
+      std::size_t pnpIdx = 0;
+      for (std::size_t jdx = 0; jdx < matches.size(); ++jdx)
+      {
+        const auto &match = matches[jdx];
+        const auto &pMp = allMapPoints[match.trainIdx];
+        const auto &ORBPoint = allORBPoints[match.queryIdx];
+        if (!pMp || pMp->isBad())
+          continue;
+        ORBPoints.push_back(ORBPoint);
+        mapPoints.push_back(pMp->getPos().clone());
+        relocBowParam.mvPnPId2MatchID[idx][pnpIdx] = jdx;
+        ++pnpIdx;
+      }
+      relocBowParam.mvpSolvers[idx] = PnPSolver::create(mapPoints, ORBPoints);
+      ++nCandidates;
+    }
+  }
+  return nCandidates;
 }
 
 /**
@@ -441,22 +497,23 @@ int Tracking::filterKFByBow(RelocBowParam &relocBowParam, std::vector<bool> &vbD
  * @param idx               输入的候选关键帧的索引
  * @param vpCandidateKFs    输入的候选关键帧
  */
-void Tracking::setCurrFrameAttrib(const std::vector<std::size_t> &vInliers, const RelocBowParam &relocBowParam,
-                                  const std::size_t &idx, std::vector<KeyFrame::SharedPtr> &vpCandidateKFs,
-                                  const cv::Mat &Rcw, const cv::Mat &tcw) {
-    mpCurrFrame->setMapPointsNull();
-    for (const std::size_t &inlierIdx : vInliers) {
-        std::size_t matchID = relocBowParam.mvPnPId2MatchID[idx].at(inlierIdx);
-        int pMpId = relocBowParam.mvAllMatches[idx][matchID].trainIdx;
-        int kPId = relocBowParam.mvAllMatches[idx][matchID].queryIdx;
-        MapPoint::SharedPtr pMp = vpCandidateKFs[idx]->getMapPoints()[pMpId];
-        if (pMp && !pMp->isBad())
-            mpCurrFrame->setMapPoint(kPId, pMp);
-    }
-    cv::Mat Tcw = cv::Mat::eye(4, 4, CV_32F);
-    Rcw.copyTo(Tcw(cv::Range(0, 3), cv::Range(0, 3)));
-    tcw.copyTo(Tcw(cv::Range(0, 3), cv::Range(3, 4)));
-    mpCurrFrame->setPose(Tcw);
+void Tracking::setCurrFrameAttrib(const std::vector<std::size_t> &vInliers, const RelocBowParam &relocBowParam, const std::size_t &idx,
+                                  std::vector<KeyFrame::SharedPtr> &vpCandidateKFs, const cv::Mat &Rcw, const cv::Mat &tcw)
+{
+  mpCurrFrame->setMapPointsNull();
+  for (const std::size_t &inlierIdx : vInliers)
+  {
+    std::size_t matchID = relocBowParam.mvPnPId2MatchID[idx].at(inlierIdx);
+    int pMpId = relocBowParam.mvAllMatches[idx][matchID].trainIdx;
+    int kPId = relocBowParam.mvAllMatches[idx][matchID].queryIdx;
+    MapPoint::SharedPtr pMp = vpCandidateKFs[idx]->getMapPoints()[pMpId];
+    if (pMp && !pMp->isBad())
+      mpCurrFrame->setMapPoint(kPId, pMp);
+  }
+  cv::Mat Tcw = cv::Mat::eye(4, 4, CV_32F);
+  Rcw.copyTo(Tcw(cv::Range(0, 3), cv::Range(0, 3)));
+  tcw.copyTo(Tcw(cv::Range(0, 3), cv::Range(3, 4)));
+  mpCurrFrame->setPose(Tcw);
 }
 
 /**
@@ -471,64 +528,73 @@ void Tracking::setCurrFrameAttrib(const std::vector<std::size_t> &vInliers, cons
  * @return true     重定位跟踪成功
  * @return false    重定位跟踪失败
  */
-bool Tracking::trackReLocalize() {
-    /// 1. 使用关键帧数据库，找到初步的候选关键帧
-    int candidateNum;
-    std::vector<KeyFrame::SharedPtr> vpCandidateKFs;
-    if (!findInitialKF(vpCandidateKFs, candidateNum))
-        return false;
+bool Tracking::trackReLocalize()
+{
+  /// 1. 使用关键帧数据库，找到初步的候选关键帧
+  int candidateNum;
+  std::vector<KeyFrame::SharedPtr> vpCandidateKFs;
+  if (!findInitialKF(vpCandidateKFs, candidateNum))
+    return false;
 
-    /// 2. 使用词袋匹配进一步筛选候选关键帧
-    std::vector<bool> vbDiscard(candidateNum, false);
-    RelocBowParam relocBowParam(candidateNum);
-    int nCandidates = filterKFByBow(relocBowParam, vbDiscard, candidateNum, vpCandidateKFs);
+  /// 2. 使用词袋匹配进一步筛选候选关键帧
+  std::vector<bool> vbDiscard(candidateNum, false);
+  RelocBowParam relocBowParam(candidateNum);
+  int nCandidates = filterKFByBow(relocBowParam, vbDiscard, candidateNum, vpCandidateKFs);
 
-    /// 3. 使用RANSAC和EPnP获取当前帧的初步位姿和内点分布
-    std::vector<std::size_t> vInliers;
-    bool bContinue = true;
-    KeyFrame::SharedPtr pRelocKf = nullptr;
-    while (nCandidates && bContinue) {
-        for (std::size_t idx = 0; idx < candidateNum; ++idx) {
-            if (vbDiscard[idx])
-                continue;
-            bool bNoMore = false;
-            PnPRet modelReti;
-            vInliers.clear();
-            bool ret = relocBowParam.mvpSolvers[idx]->iterate(5, modelReti, bNoMore, vInliers);
-            if (bNoMore) {
-                vbDiscard[idx] = true;
-                --nCandidates;
-            }
-            if (!ret)
-                continue;
-            setCurrFrameAttrib(vInliers, relocBowParam, idx, vpCandidateKFs, modelReti.mRcw, modelReti.mtcw);
-            int nInliers = Optimizer::OptimizePoseOnly(mpCurrFrame);
-            if (nInliers < 10)
-                continue;
-            else if (nInliers < 50) {
-                /// 4. 使用重投影匹配进行精确匹配
-                if (addMatchByProject(vpCandidateKFs[idx], nInliers)) {
-                    pRelocKf = vpCandidateKFs[idx];
-                    bContinue = false;
-                    break;
-                }
-            } else {
-                pRelocKf = vpCandidateKFs[idx];
-                bContinue = false;
-                break;
-            }
+  /// 3. 使用RANSAC和EPnP获取当前帧的初步位姿和内点分布
+  std::vector<std::size_t> vInliers;
+  bool bContinue = true;
+  KeyFrame::SharedPtr pRelocKf = nullptr;
+  while (nCandidates && bContinue)
+  {
+    for (std::size_t idx = 0; idx < candidateNum; ++idx)
+    {
+      if (vbDiscard[idx])
+        continue;
+      bool bNoMore = false;
+      PnPRet modelReti;
+      vInliers.clear();
+      bool ret = relocBowParam.mvpSolvers[idx]->iterate(5, modelReti, bNoMore, vInliers);
+      if (bNoMore)
+      {
+        vbDiscard[idx] = true;
+        --nCandidates;
+      }
+      if (!ret)
+        continue;
+      setCurrFrameAttrib(vInliers, relocBowParam, idx, vpCandidateKFs, modelReti.mRcw, modelReti.mtcw);
+      int nInliers = Optimizer::OptimizePoseOnly(mpCurrFrame);
+      if (nInliers < 10)
+        continue;
+      else if (nInliers < 50)
+      {
+        /// 4. 使用重投影匹配进行精确匹配
+        if (addMatchByProject(vpCandidateKFs[idx], nInliers))
+        {
+          pRelocKf = vpCandidateKFs[idx];
+          bContinue = false;
+          break;
         }
+      }
+      else
+      {
+        pRelocKf = vpCandidateKFs[idx];
+        bContinue = false;
+        break;
+      }
     }
-    if (bContinue)
-        return false;
-    if (pRelocKf && !pRelocKf->isBad()) {
-        pRelocKf->setNotErased(true);
-        if (mpRefKf && !mpRefKf->isBad())
-            mpRefKf->setNotErased(false);
-        mpRefKf = pRelocKf;
-    }
-    mnLastRelocId = mpCurrFrame->getID();
-    return true;
+  }
+  if (bContinue)
+    return false;
+  if (pRelocKf && !pRelocKf->isBad())
+  {
+    pRelocKf->setNotErased(true);
+    if (mpRefKf && !mpRefKf->isBad())
+      mpRefKf->setNotErased(false);
+    mpRefKf = pRelocKf;
+  }
+  mnLastRelocId = mpCurrFrame->getID();
+  return true;
 }
 
 /**
@@ -543,21 +609,23 @@ bool Tracking::trackReLocalize() {
  * @return true     成功
  * @return false    失败
  */
-bool Tracking::addMatchByProject(KeyFrame::SharedPtr pKFrame, int &nInliers) {
-    if (!pKFrame || pKFrame->isBad())
-        return false;
-    ORBMatcher matcher2(0.9, true);
-    std::vector<cv::DMatch> matches2;
-    int nAddition = matcher2.searchByProjection(mpCurrFrame, pKFrame, matches2, 10);
+bool Tracking::addMatchByProject(KeyFrame::SharedPtr pKFrame, int &nInliers)
+{
+  if (!pKFrame || pKFrame->isBad())
+    return false;
+  ORBMatcher matcher2(0.9, true);
+  std::vector<cv::DMatch> matches2;
+  int nAddition = matcher2.searchByProjection(mpCurrFrame, pKFrame, matches2, 10);
+  if (nAddition + nInliers < 50)
+    return false;
+  nInliers = Optimizer::OptimizePoseOnly(mpCurrFrame);
+  if (nInliers < 50)
+  {
+    nAddition = matcher2.searchByProjection(mpCurrFrame, pKFrame, matches2, 3);
     if (nAddition + nInliers < 50)
-        return false;
-    nInliers = Optimizer::OptimizePoseOnly(mpCurrFrame);
-    if (nInliers < 50) {
-        nAddition = matcher2.searchByProjection(mpCurrFrame, pKFrame, matches2, 3);
-        if (nAddition + nInliers < 50)
-            return false;
-    }
-    return true;
+      return false;
+  }
+  return true;
 }
 
 /**
@@ -570,38 +638,40 @@ bool Tracking::addMatchByProject(KeyFrame::SharedPtr pKFrame, int &nInliers) {
  * @return true
  * @return false
  */
-bool Tracking::trackLocalMap() {
-    buildLocalMap();
-    ORBMatcher matcher(0.8, true);
-    float th = 3;
-    if (mnLastRelocId > 0)
-        if (mpCurrFrame->getID() < mnLastRelocId + 2)
-            th = 5;
+bool Tracking::trackLocalMap()
+{
+  buildLocalMap();
+  ORBMatcher matcher(0.8, true);
+  float th = 3;
+  if (mnLastRelocId > 0)
+    if (mpCurrFrame->getID() < mnLastRelocId + 2)
+      th = 5;
 
-    std::vector<cv::DMatch> matches;
-    int nMatches = 0;
-    {
-        std::unique_lock<std::mutex> lock(mMutexLMps);
-        nMatches = matcher.searchByProjection(mpCurrFrame, mvpLocalMps, th, matches);
-    }
-    if (nMatches < 30)
-        return false;
-    Optimizer::OptimizePoseOnly(mpCurrFrame);
-    int nInliers = 0;
-    std::size_t nNum = mpCurrFrame->getLeftKeyPoints().size();
-    auto mps = mpCurrFrame->getMapPoints();
-    for (std::size_t idx = 0; idx < nNum; ++idx) {
-        auto &pMp = mps[idx];
-        if (pMp && pMp->isInMap())
-            ++nInliers;
-        else if (pMp && !pMp->isInMap())
-            mpCurrFrame->setMapPoint(idx, nullptr);
-    }
-    if (nInliers < 30)
-        return false;
-    if (mpCurrFrame->getID() < mnLastRelocId + mnMaxFrames && nInliers < 50)
-        return false;
-    return true;
+  std::vector<cv::DMatch> matches;
+  int nMatches = 0;
+  {
+    std::unique_lock<std::mutex> lock(mMutexLMps);
+    nMatches = matcher.searchByProjection(mpCurrFrame, mvpLocalMps, th, matches);
+  }
+  if (nMatches < 30)
+    return false;
+  Optimizer::OptimizePoseOnly(mpCurrFrame);
+  int nInliers = 0;
+  std::size_t nNum = mpCurrFrame->getLeftKeyPoints().size();
+  auto mps = mpCurrFrame->getMapPoints();
+  for (std::size_t idx = 0; idx < nNum; ++idx)
+  {
+    auto &pMp = mps[idx];
+    if (pMp && pMp->isInMap())
+      ++nInliers;
+    else if (pMp && !pMp->isInMap())
+      mpCurrFrame->setMapPoint(idx, nullptr);
+  }
+  if (nInliers < 30)
+    return false;
+  if (mpCurrFrame->getID() < mnLastRelocId + mnMaxFrames && nInliers < 50)
+    return false;
+  return true;
 }
 
 /**
@@ -612,14 +682,15 @@ bool Tracking::trackLocalMap() {
  *      3. 使用Tracking进行临时地图点的维护
  *      4. 值得注意的是，维护的临时地图点中，有nullptr
  */
-void Tracking::processLastFrame() {
-    auto refKf = mpLastFrame->getRefKF();
-    assert(refKf && "上一帧的参考关键帧为空！");
-    mpLastFrame->setPose(mTlr * refKf->getPose());
+void Tracking::processLastFrame()
+{
+  auto refKf = mpLastFrame->getRefKF();
+  assert(refKf && "上一帧的参考关键帧为空！");
+  mpLastFrame->setPose(mTlr * refKf->getPose());
 
-    std::vector<MapPoint::SharedPtr> mapPoints;
-    mpLastFrame->unProject(mapPoints);
-    std::swap(mapPoints, mvpTempMappoints);
+  std::vector<MapPoint::SharedPtr> mapPoints;
+  mpLastFrame->unProject(mapPoints);
+  std::swap(mapPoints, mvpTempMappoints);
 }
 
 /**
@@ -647,79 +718,89 @@ void Tracking::processLastFrame() {
  * @return true     局部建图线程需要插入关键帧
  * @return false    局部建图线程不需要插入关键帧
  */
-bool Tracking::needNewKeyFrame() {
-    if (mbOnlyTracking)
-        return false;
+bool Tracking::needNewKeyFrame()
+{
+  if (mbOnlyTracking)
+    return false;
 
-    if (mnLastRelocId > 0)
-        if (mpCurrFrame->getID() - mnLastRelocId <= mnMaxFrames)
-            return false;
-    if (mpLocalMapper->isRequestStop()) {
-        return false;
+  if (mnLastRelocId > 0)
+    if (mpCurrFrame->getID() - mnLastRelocId <= mnMaxFrames)
+      return false;
+  if (mpLocalMapper->isRequestStop())
+  {
+    return false;
+  }
+
+  auto refMps = mpRefKf->getMapPoints();
+  auto currMps = mpCurrFrame->getMapPoints();
+
+  int nRefMps = 0, nCurrMps = 0;
+  int nObs = mpRefKf->getID() == 0 ? 0 : 1;
+  for (std::size_t idx = 0; idx < refMps.size(); ++idx)
+  {
+    auto &pMp = refMps[idx];
+    if (pMp && !pMp->isBad() && pMp->getObsNum() > nObs)
+      ++nRefMps;
+    else
+      mpRefKf->setMapPoint(idx, nullptr);
+  }
+  for (std::size_t idx = 0; idx < currMps.size(); ++idx)
+  {
+    auto &pMp = currMps[idx];
+    if (pMp && !pMp->isBad() && pMp->isInMap())
+      ++nCurrMps;
+    else
+    {
+      mpCurrFrame->setMapPoint(idx, nullptr);
+      currMps[idx] = nullptr;
     }
+  }
+  auto &vDepths = mpCurrFrame->getDepth();
+  double depthTh = Camera::mfBl * mnDepthTh;
 
-    auto refMps = mpRefKf->getMapPoints();
-    auto currMps = mpCurrFrame->getMapPoints();
-
-    int nRefMps = 0, nCurrMps = 0;
-    int nObs = mpRefKf->getID() == 0 ? 0 : 1;
-    for (std::size_t idx = 0; idx < refMps.size(); ++idx) {
-        auto &pMp = refMps[idx];
-        if (pMp && !pMp->isBad() && pMp->getObsNum() > nObs)
-            ++nRefMps;
-        else
-            mpRefKf->setMapPoint(idx, nullptr);
+  int nTrackedClose = 0, nNoTrackedClose = 0;
+  for (std::size_t idx = 0; idx < vDepths.size(); ++idx)
+  {
+    const double &depth = vDepths[idx];
+    auto &pMp = currMps[idx];
+    if (depth < depthTh && depth > 0)
+    {
+      if (pMp && !pMp->isBad())
+        ++nTrackedClose;
+      else
+        ++nNoTrackedClose;
     }
-    for (std::size_t idx = 0; idx < currMps.size(); ++idx) {
-        auto &pMp = currMps[idx];
-        if (pMp && !pMp->isBad() && pMp->isInMap())
-            ++nCurrMps;
-        else {
-            mpCurrFrame->setMapPoint(idx, nullptr);
-            currMps[idx] = nullptr;
-        }
-    }
-    auto &vDepths = mpCurrFrame->getDepth();
-    double depthTh = Camera::mfBl * mnDepthTh;
+  }
+  float ratio = nCurrMps / ((float)nRefMps + 1e-5);
 
-    int nTrackedClose = 0, nNoTrackedClose = 0;
-    for (std::size_t idx = 0; idx < vDepths.size(); ++idx) {
-        const double &depth = vDepths[idx];
-        auto &pMp = currMps[idx];
-        if (depth < depthTh && depth > 0) {
-            if (pMp && !pMp->isBad())
-                ++nTrackedClose;
-            else
-                ++nNoTrackedClose;
-        }
-    }
-    float ratio = nCurrMps / ((float)nRefMps + 1e-5);
+  bool bNeedClose = nTrackedClose < 100 && nNoTrackedClose > 70;
+  bool bLocalMapperIdle = mpLocalMapper->getAcceptKF();
+  bool c1a = mpCurrFrame->getID() - mnLastInsertId > mnMaxFrames;
+  bool c1b = mpCurrFrame->getID() - mnLastInsertId > mnMinFrames && bLocalMapperIdle;
+  bool c1c = ratio < 0.25 || bNeedClose;
+  bool c1 = c1a || c1b || c1c;
 
-    bool bNeedClose = nTrackedClose < 100 && nNoTrackedClose > 70;
-    bool bLocalMapperIdle = mpLocalMapper->getAcceptKF();
-    bool c1a = mpCurrFrame->getID() - mnLastInsertId > mnMaxFrames;
-    bool c1b = mpCurrFrame->getID() - mnLastInsertId > mnMinFrames && bLocalMapperIdle;
-    bool c1c = ratio < 0.25 || bNeedClose;
-    bool c1 = c1a || c1b || c1c;
-
-    float ratioTh = 0.75;
-    if (mpMap->keyFramesInMap() < 2)
-        ratioTh = 0.4;
-    bool c2a = ratio < ratioTh;
-    bool c2b = bNeedClose;
-    bool c2 = c2a || c2b;
-    bool flag1 = c1 && c2;
-    if (!flag1)
-        return false;
-    if (bLocalMapperIdle)
-        return true;
-    else {
-        mpLocalMapper->setAbortBA(true);
-        if (mpLocalMapper->getKFNum() < 3) {
-            return true;
-        } else
-            return false;
+  float ratioTh = 0.75;
+  if (mpMap->keyFramesInMap() < 2)
+    ratioTh = 0.4;
+  bool c2a = ratio < ratioTh;
+  bool c2b = bNeedClose;
+  bool c2 = c2a || c2b;
+  bool flag1 = c1 && c2;
+  if (!flag1)
+    return false;
+  if (bLocalMapperIdle)
+    return true;
+  else
+  {
+    mpLocalMapper->setAbortBA(true);
+    if (mpLocalMapper->getKFNum() < 3)
+    {
+      return true;
     }
+    else
+      return false;
+  }
 }
 
 } // namespace ORB_SLAM2_ROS2
